services/ocr-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
Pillow==10.1.0
pydantic==2.4.2
paddleocr==2.7.3
paddlepaddle==2.6.0
numpy==1.24.3
opencv-python-headless==4.8.1.78
requests==2.31.0
jsondiff==2.0.0
scikit-image==0.22.0
lpips==0.1.4
torch==2.1.0
torchvision==0.16.0
